completing assignment
