+++ src/test/java/guru/springframework/sfgpetclinic/model/PersonRepeatedTests.java
@@ -1,0 +1,31 @@
+package guru.springframework.sfgpetclinic.model;
+
+import guru.springframework.sfgpetclinic.ModelRepeatedTests;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.RepeatedTest;
+import org.junit.jupiter.api.RepetitionInfo;
+import org.junit.jupiter.api.TestInfo;
+
+/**
+ * Created by jt on 2018-10-28.
+ */
+public class PersonRepeatedTests implements ModelRepeatedTests {
+
+    @RepeatedTest(value = 10, name = "{displayName} : {currentRepetition} - {totalRepetitions}")
+    @DisplayName("My Repeated Test")
+    void myRepeatedTest() {
+        //todo - impl
+    }
+
+    @RepeatedTest(5)
+    void myRepeatedTestWithDI(TestInfo testInfo, RepetitionInfo repetitionInfo) {
+        System.out.println(testInfo.getDisplayName() + ": " + repetitionInfo.getCurrentRepetition());
+
+    }
+
+    @RepeatedTest(value = 5, name = "{displayName} : {currentRepetition} | {totalRepetitions}")
+    @DisplayName("My Assignment Repeated Test")
+    void myAssignmentRepeated() {
+        //todo impl
+    }
+}
